prom

+++ app/controllers/admin/common_controller.rb
@@ -1,5 +1,7 @@
 module Admin
   class CommonController < AdminController
+    before_action :current_company
+
     def index
       @models = model.columns_hash["company_id"].present? ? model.where(company_id: current_company.id) : model.all
     end
@@ -15,21 +17,21 @@ module Admin
     end
 
     def show
-      @model = find_model
+      @model = find_model if ((find_model.company_id == current_company.id) rescue true )
     end
 
     def edit
       @model = find_model
-      render_if_json
+      ((@model.company_id == current_company.id) rescue true ) ? render_if_json : (render text: "Страница не найдена 404")
     end
 
     def update
-      find_model.update(params_model)
+      find_model.update(params_model) if ((find_model.company_id == current_company.id) rescue true )
       redirect_to_index
     end
 
     def remove
-      find_model.destroy
+      find_model.destroy if ((find_model.company_id == current_company.id) rescue true )
       redirect_to_index
     end
 
@@ -51,7 +53,11 @@ module Admin
     end
 
     def find_model
-      model.find(params[:id])
+      begin 
+        model.find(params[:id])
+      rescue
+        redirect_to "/404"
+      end
     end
 
     def params_model
